Multiple changes to try to understand why BBM is behaving strangely
This is not final by any stretch, but rather a check-point along the
way. Importantly:
 - We need to advect the stress tensor. It's done here, but maybe not
   correctly
 - td needs to depend on the grid size

Problems remaining:
 - Strange features where the ice is "stuck"
 - Results are time step dependent
 - This clearly still looks "wrong"

+++ dynamics/src/include/MEBParameters.hpp
@@ -33,7 +33,7 @@ public:
     double P0; //! < \param Constant to parametrize Pmax
     double damage_timescale = 1.; //<! Damage timescale
     double undamaged_time_relaxation_sigma = 1e5; //!< \param lambda
-    double exponent_relaxation_sigma;
+    int exponent_relaxation_sigma;
     double c0; //! \param
     double sigma_c0; //! \param
